Trainwreck_ImageNet-100/requirements.txt
cleverhans==4.0.0
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.1.0
scikit_learn==1.6.1
scipy==1.15.2
timm==1.0.15
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.64.1
